Remove obsolete SunPRO workaround, and workaround warning.

+++ Straight_skeleton_2/include/CGAL/constructions/Straight_skeleton_cons_ftC2.h
@@ -50,7 +50,7 @@ inline CORE::BigFloat MP_Float_to_BigFloat( MP_Float const& b )
   }
 
   // The cast is necessary for SunPro.
-  return d * CORE::BigFloat::exp2(static_cast<int>(exp * log_limb));
+  return d * CORE::BigFloat::exp2(exp * log_limb);
 }
 #endif
